pydantic-ai/requirements.txt
pydantic-ai-slim[openai]
bpython
